requirements.txt
aiogram==2.25.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
aiohttp<3.9.0,>=3.8.0
python-dotenv==1.0.0
APScheduler==3.10.4